feat(agents): negotiate markdown on the homepage, advertise it in Link and robots
The Cloudflare agent-ready scan scored career-ops.org 0/1 on content
negotiation. The check was right and our earlier verification was wrong: it
probes the HOMEPAGE, and src/proxy.ts only matched /docs. So an agent landing
on career-ops.org (the URL in every share, badge and press mention) asked for
markdown and got HTML, with no way to guess the whole agent-facing layer lived
under /docs.

Three changes, one pass, per search-ops' 2026-08-10 verdict:

1. `/` with `Accept: text/markdown` now serves the canonical llms.txt body as
   text/markdown. The home is marketing; its correct agent representation IS
   the index, so there is one artifact to maintain rather than a parallel
   home.md. The llms.txt builder moves to src/lib/llms-index.ts so the route
   and the markdown mirror share the same bytes.

   NOT implemented in the proxy: verified that the proxy does not run for `/`
   at all (a debug header set unconditionally on the root never appeared), the
   same class of Next 16 surprise as the `.md` interception in July. It is a
   next.config `beforeFiles` rewrite gated on `has`/`missing` Accept instead,
   which fires deterministically at the routing layer.

2. `Link:` headers (RFC 8288) on `/` and `/docs**` advertising llms.txt,
   llms-full.txt and the markdown alternate, so a client issuing only HEAD
   discovers the agent layer without parsing HTML.

3. `Content-Signal: search=yes, ai-input=yes, ai-train=yes` in robots.txt.
   This is deliberately the opposite of the CDN default (`ai-train=no`): the
   project wants to be read, cited AND trained on, and an explicit declaration
   protects against a CDN default change quietly dropping the site out of the
   next training crawl. Ratified by search-ops, FYI to venture-ops. robots
   becomes a route handler because Next's metadata helper cannot emit
   non-standard directives.

Verified locally on a production build: agent gets markdown, browser and a
bare curl still get HTML, /docs unchanged, robots carries the signal on all 16
user-agent blocks.

Not in this PR: real .md for /manifesto, /methodology and /press. Those pages
are TSX with no markdown source, so serving them needs a design decision about
where that source lives, not a router change. Raised separately.
